+++ request_api_helper/lib/request_api_helper.dart
@@ -2,3 +2,8 @@ library request_api_helper;
 
 export 'package:fluttertoast/fluttertoast.dart';
 export 'serverSwitcher.dart';
+
+export 'package:request_api_helper/model/config_model.dart';
+export 'package:request_api_helper/model/redirect_helper.dart';
+export 'session.dart';
+export 'background.dart';