+++ request_checker/lib/background.dart
@@ -1,27 +1,28 @@
-
 import 'package:flutter/material.dart';
+import 'model/redirect_helper.dart';
 
-bool isLoading = false ;
+bool isLoading = false;
 
-class BackDark{
-  dynamic view;
-  BackDark({
-    this.view
-  });
+class BackDark {
+  Redirects? view;
+  bool? barrierDismissible;
+  BackDark({this.view, this.barrierDismissible});
   dialog(context) async {
-    return showDialog<void>(
-      context: context,
-      barrierDismissible: false, // user must tap button!
-      builder: (BuildContext context) {
-        return WillPopScope(
-          onWillPop: () async {
-            return !isLoading;
-          },
-          child: Center(
-            child: view != null && view != true ? view : RefreshProgressIndicator(),
-          ),
-        );
-      },
-    );
+    Future.delayed(Duration.zero, () {
+      return showDialog<void>(
+        context: context,
+        barrierDismissible: barrierDismissible ?? false, // user must tap button!
+        builder: (BuildContext context) {
+          return WillPopScope(
+            onWillPop: () async {
+              return !isLoading;
+            },
+            child: Center(
+              child: view!.widget ?? RefreshProgressIndicator(),
+            ),
+          );
+        },
+      );
+    });
   }
 }