+++ request_checker/lib/model/config_model.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/material.dart' show BuildContext;
-import 'package:request_api_helper/model/redirect_helper.dart';
+import 'package:request_checker/model/redirect_helper.dart';
 
 enum RESTAPI { POST, GET, PUT, DELETE, PATCH }
 BuildContext? currentContext;