+++ request_checker/lib/model/redirect_helper.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart' show Widget;
+
+class Redirects {
+  /// if redirect with custom widget
+  Widget? widget;
+
+  /// if redirect is active
+  bool? toogle;
+
+  Redirects({this.toogle, this.widget});
+}