+++ request_checker/lib/request.dart
@@ -1,6 +1,6 @@
-import 'package:request_api_helper/model/config_model.dart';
+import 'package:request_checker/model/config_model.dart';
 import 'package:flutter/material.dart' show BuildContext;
-import 'package:request_api_helper/rest_api.dart' show OnUploadProgressCallback;
+import 'package:request_checker/rest_api.dart' show OnUploadProgressCallback;
 import 'base_request.dart';
 
 Future send({CustomRequestData? customData, RequestData? data, String? name, RequestApiHelperConfigData? changeConfig, BuildContext? context, required RESTAPI type, bool? singleContext, OnUploadProgressCallback? onUploadProgress}) => _withClient(customData, data, name, changeConfig, context, type, singleContext, onUploadProgress);